Add CreateElementColumn  component
